+++ build.rs
@@ -1,9 +1,11 @@
 use rustc_version::{version_meta, Channel};
 
 fn main() {
-    let has_generic_const_exprs = version_meta().unwrap().channel == Channel::Nightly;
-    if has_generic_const_exprs {
+    // HACK: Ideally we would test for features instead of release channel.
+    let nightly = version_meta().unwrap().channel == Channel::Nightly;
+    if nightly {
         println!("cargo:rustc-cfg=has_generic_const_exprs");
+        println!("cargo:rustc-cfg=has_cfg_doc");
     }
     println!("cargo:rerun-if-changed=build.rs");
 }