+++ src/modular.rs
@@ -208,6 +208,7 @@ mod tests {
 }
 
 #[cfg(feature = "bench")]
+#[doc(hidden)]
 pub mod bench {
     use super::*;
     use crate::{const_for, nlimbs};