+++ src/support/quickcheck.rs
@@ -1,4 +1,7 @@
+//! Support for the [`quickcheck`](https://crates.io/crates/quickcheck) crate.
 #![cfg(feature = "quickcheck")]
+#![cfg_attr(has_doc_cfg, doc(cfg(feature = "quickcheck")))]
+
 use crate::Uint;
 use quickcheck::{Arbitrary, Gen};