+++ src/support/rand.rs
@@ -1,6 +1,9 @@
+//! Support for the [`rand`](https://crates.io/crates/rand) crate.
+#![cfg(feature = "rand")]
+#![cfg_attr(has_doc_cfg, doc(cfg(feature = "rand")))]
+
 // FEATURE: Implement the Uniform distribution.
 
-#![cfg(feature = "rand")]
 use crate::Uint;
 use rand::{
     distributions::{Distribution, Standard, Uniform},