+++ src/add.rs
@@ -267,6 +267,7 @@ pub mod tests {
 }
 
 #[cfg(feature = "bench")]
+#[doc(hidden)]
 pub mod bench {
     use super::*;
     use crate::{const_for, nlimbs};