+++ src/algorithms/gcd/mod.rs
@@ -243,6 +243,7 @@ mod tests {
 }
 
 #[cfg(feature = "bench")]
+#[doc(hidden)]
 pub mod bench {
     use super::*;
     use criterion::Criterion;