+++ src/div.rs
@@ -136,6 +136,7 @@ pub mod tests {
 }
 
 #[cfg(feature = "bench")]
+#[doc(hidden)]
 pub mod bench {
     use super::*;
     use crate::{const_for, nlimbs};