+++ src/lib.rs
@@ -15,6 +15,8 @@
 )]
 // See <https://github.com/taiki-e/coverage-helper>
 #![cfg_attr(coverage_nightly, feature(no_coverage))]
+// See <https://stackoverflow.com/questions/61417452/how-to-get-a-feature-requirement-tag-in-the-documentation-generated-by-cargo-do>
+#![cfg_attr(has_doc_cfg, feature(doc_cfg))]
 
 mod add;
 pub mod algorithms;
@@ -305,6 +307,7 @@ mod test {
 }
 
 #[cfg(feature = "bench")]
+#[doc(hidden)]
 pub mod bench {
     use super::*;
     use criterion::Criterion;