+++ Changelog.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `reduce_mod`, `add_mod`, `mul_mod`, `pow_mod`.
+- Added `reduce_mod`, `add_mod`, `mul_mod`, `pow_mod`, `inv_mod`.
+- Added `gcd`, `gcd_extended`, `lcm`.
 - Added `num-bigint` and `ark-ff` support.
 - Made `algorithms` public, but with unstable API for now.