+++ Changelog.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `inv_mod`, `gcd`, `gcd_extended`, `lcm`.
+- Added `inv_mod`, `mul_redc`, `gcd`, `gcd_extended`, `lcm`.
 
 ## [1.2.0] — 2022-06-03