+++ Changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `inv_mod`, `gcd`, `gcd_extended`, `lcm`.
+- Added `sqlx` support.
 
 ## [1.2.0] — 2022-06-03